+++ src/main/java/duke/Parser.java
@@ -1,3 +1,10 @@
+package duke;
+
+import duke.command.*;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Todo;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
@@ -46,7 +53,7 @@ public class Parser {
             if (description.equals("")) {
                 return new ListCommand();
             } else {
-                throw new DukeException("Command 'list' does not need additional arguments");
+                throw new DukeException("duke.command.Command 'list' does not need additional arguments");
             }
         case ("done"):
             try {
@@ -91,7 +98,7 @@ public class Parser {
                 throw new DukeException("Type 'bye' only to exit");
             }
         default:
-            throw new DukeException("Command not recognised!");
+            throw new DukeException("duke.command.Command not recognised!");
         }
     }
 }
