+++ src/main/java/duke/Ui.java
@@ -1,4 +1,5 @@
-import java.time.LocalDateTime;
+package duke;
+
 import java.util.Scanner;
 
 public class Ui {
