+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,20 +1,27 @@
-public class DoneCommand extends Command {
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class DeleteCommand extends Command {
     private int index;
 
-    public DoneCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
-
     @Override
     public void execute(TaskList taskList, Storage storage, Ui ui) {
         try {
-            String doneString = taskList.markAsDone(index);
+            String taskString = taskList.deleteTask(index);
             storage.rewrite(taskList);
-            ui.showDoneMessage(doneString);
+            ui.showDeleteMessage(taskString, taskList);
         } catch (IndexOutOfBoundsException e) {
             throw new DukeException("Invalid task number!");
         }
     }
+
     @Override
     public boolean isExit() {
         return false;
